keep coverage diagnostics within size gate
